refactor(cmd/bd): extract business logic from dep.go and swarm.go into internal/ packages (bd-za9)
Move pure business logic out of cmd/bd/ into internal/deps/ and internal/swarm/
so that cmd/bd merely wires up cobra commands with thin delegation wrappers.

internal/deps/ (new):
- IsChildOf, ValidateExternalRef, IsExternalRef, ParseExternalRef
- FilterTreeByStatus, MergeBidirectionalTrees, GetStatusEmoji
- OutputMermaidTree, FormatTreeNode, TreeRenderer/RenderTree

internal/swarm/ (new):
- Analysis, ReadyFront, IssueNode, Storage interface, Status, StatusIssue types
- FindExistingSwarm, GetEpicChildren, AnalyzeEpicForSwarm
- DetectStructuralIssues, ComputeReadyFronts, GetSwarmStatus

cmd/bd/dep.go and cmd/bd/swarm.go now contain only cobra command definitions,
CLI-specific wiring (globals, output formatting), and thin wrappers that delegate
to the new packages. No test files were modified; all existing tests pass.
