+++ cmd/bd/swarm.go
@@ -11,10 +11,19 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/steveyegge/beads/internal/storage"
+	"github.com/steveyegge/beads/internal/swarm"
 	"github.com/steveyegge/beads/internal/types"
 	"github.com/steveyegge/beads/internal/ui"
 	"github.com/steveyegge/beads/internal/utils"
 )
+
+// Type aliases for backward compatibility with tests and other files in cmd/bd.
+type SwarmAnalysis = swarm.Analysis
+type ReadyFront = swarm.ReadyFront
+type IssueNode = swarm.IssueNode
+type SwarmStorage = swarm.Storage
+type SwarmStatus = swarm.Status
+type StatusIssue = swarm.StatusIssue
 
 var swarmCmd = &cobra.Command{
 	Use:     "swarm",
@@ -26,110 +35,24 @@ A swarm is a structured body of work defined by an epic and its children,
 with dependencies forming a DAG (directed acyclic graph) of work.`,
 }
 
-// SwarmAnalysis holds the results of analyzing an epic's structure for swarming.
-type SwarmAnalysis struct {
-	EpicID            string                `json:"epic_id"`
-	EpicTitle         string                `json:"epic_title"`
-	TotalIssues       int                   `json:"total_issues"`
-	ClosedIssues      int                   `json:"closed_issues"`
-	ReadyFronts       []ReadyFront          `json:"ready_fronts"`
-	MaxParallelism    int                   `json:"max_parallelism"`
-	EstimatedSessions int                   `json:"estimated_sessions"`
-	Warnings          []string              `json:"warnings"`
-	Errors            []string              `json:"errors"`
-	Swarmable         bool                  `json:"swarmable"`
-	Issues            map[string]*IssueNode `json:"issues,omitempty"` // Only included with --verbose
-}
-
-// ReadyFront represents a group of issues that can be worked on in parallel.
-type ReadyFront struct {
-	Wave   int      `json:"wave"`
-	Issues []string `json:"issues"`
-	Titles []string `json:"titles,omitempty"` // Only for human output
-}
-
-// IssueNode represents an issue in the dependency graph.
-type IssueNode struct {
-	ID           string   `json:"id"`
-	Title        string   `json:"title"`
-	Status       string   `json:"status"`
-	Priority     int      `json:"priority"`
-	DependsOn    []string `json:"depends_on"`     // What this issue depends on
-	DependedOnBy []string `json:"depended_on_by"` // What depends on this issue
-	Wave         int      `json:"wave"`           // Which ready front this belongs to (-1 if blocked by cycle)
-}
-
-// SwarmStorage defines the storage interface needed by swarm commands.
-type SwarmStorage interface {
-	GetIssue(context.Context, string) (*types.Issue, error)
-	GetDependents(context.Context, string) ([]*types.Issue, error)
-	GetDependencyRecords(context.Context, string) ([]*types.Dependency, error)
-}
-
-// findExistingSwarm returns the swarm molecule for an epic, if one exists.
-// Returns nil if no swarm molecule is linked to the epic.
+// findExistingSwarm delegates to swarm.FindExistingSwarm.
 func findExistingSwarm(ctx context.Context, s SwarmStorage, epicID string) (*types.Issue, error) {
-	// Get all issues that depend on the epic
-	dependents, err := s.GetDependents(ctx, epicID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get epic dependents: %w", err)
-	}
-
-	// Find a swarm molecule with relates-to dependency to this epic
-	for _, dep := range dependents {
-		// Only consider molecules (GetDependents doesn't populate mol_type, so we fetch full issue)
-		if dep.IssueType != "molecule" {
-			continue
-		}
-
-		// Get full issue to check mol_type
-		fullIssue, err := s.GetIssue(ctx, dep.ID)
-		if err != nil || fullIssue == nil {
-			continue
-		}
-		if fullIssue.MolType != types.MolTypeSwarm {
-			continue
-		}
-
-		// Verify it's linked via relates-to
-		deps, err := s.GetDependencyRecords(ctx, dep.ID)
-		if err != nil {
-			continue
-		}
-		for _, d := range deps {
-			if d.DependsOnID == epicID && d.Type == types.DepRelatesTo {
-				return fullIssue, nil
-			}
-		}
-	}
-
-	return nil, nil
+	return swarm.FindExistingSwarm(ctx, s, epicID)
 }
 
-// getEpicChildren returns all child issues of an epic (via parent-child dependencies).
+// getEpicChildren delegates to swarm.GetEpicChildren.
 func getEpicChildren(ctx context.Context, s SwarmStorage, epicID string) ([]*types.Issue, error) {
-	// Get all issues that depend on the epic
-	allDependents, err := s.GetDependents(ctx, epicID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get epic dependents: %w", err)
-	}
+	return swarm.GetEpicChildren(ctx, s, epicID)
+}
 
-	// Filter to only parent-child relationships by checking each dependent's dependency records
-	var children []*types.Issue
-	for _, dependent := range allDependents {
-		deps, err := s.GetDependencyRecords(ctx, dependent.ID)
-		if err != nil {
-			continue // Skip issues we can't query
-		}
-		for _, dep := range deps {
-			if dep.DependsOnID == epicID && dep.Type == types.DepParentChild {
-				children = append(children, dependent)
-				break
-			}
-		}
-	}
+// analyzeEpicForSwarm delegates to swarm.AnalyzeEpicForSwarm.
+func analyzeEpicForSwarm(ctx context.Context, s SwarmStorage, epic *types.Issue) (*SwarmAnalysis, error) {
+	return swarm.AnalyzeEpicForSwarm(ctx, s, epic)
+}
 
-	return children, nil
+// getSwarmStatus delegates to swarm.GetSwarmStatus.
+func getSwarmStatus(ctx context.Context, s SwarmStorage, epic *types.Issue) (*SwarmStatus, error) {
+	return swarm.GetSwarmStatus(ctx, s, epic)
 }
 
 var swarmValidateCmd = &cobra.Command{
@@ -158,18 +81,15 @@ Examples:
 		ctx := rootCtx
 		verbose, _ := cmd.Flags().GetBool("verbose")
 
-		// Swarm commands require direct store access
 		if store == nil {
 			FatalErrorRespectJSON("no database connection")
 		}
 
-		// Resolve epic ID
 		epicID, err := utils.ResolvePartialID(ctx, store, args[0])
 		if err != nil {
 			FatalErrorRespectJSON("epic '%s' not found: %v", args[0], err)
 		}
 
-		// Get the epic
 		epic, err := store.GetIssue(ctx, epicID)
 		if err != nil {
 			FatalErrorRespectJSON("failed to get epic: %v", err)
@@ -178,18 +98,15 @@ Examples:
 			FatalErrorRespectJSON("epic '%s' not found", epicID)
 		}
 
-		// Verify it's an epic
 		if epic.IssueType != types.TypeEpic && epic.IssueType != "molecule" {
 			FatalErrorRespectJSON("'%s' is not an epic or molecule (type: %s)", epicID, epic.IssueType)
 		}
 
-		// Analyze the epic structure
 		analysis, err := analyzeEpicForSwarm(ctx, store, epic)
 		if err != nil {
 			FatalErrorRespectJSON("failed to analyze epic: %v", err)
 		}
 
-		// Include detailed graph only in verbose mode
 		if !verbose {
 			analysis.Issues = nil
 		}
@@ -202,310 +119,12 @@ Examples:
 			return
 		}
 
-		// Human-readable output
 		renderSwarmAnalysis(analysis)
 
 		if !analysis.Swarmable {
 			os.Exit(1)
 		}
 	},
-}
-
-// analyzeEpicForSwarm performs structural analysis of an epic for swarm execution.
-func analyzeEpicForSwarm(ctx context.Context, s SwarmStorage, epic *types.Issue) (*SwarmAnalysis, error) {
-	analysis := &SwarmAnalysis{
-		EpicID:    epic.ID,
-		EpicTitle: epic.Title,
-		Swarmable: true,
-		Issues:    make(map[string]*IssueNode),
-	}
-
-	// Get all child issues of the epic
-	childIssues, err := getEpicChildren(ctx, s, epic.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(childIssues) == 0 {
-		analysis.Warnings = append(analysis.Warnings, "Epic has no children")
-		return analysis, nil
-	}
-
-	analysis.TotalIssues = len(childIssues)
-
-	// Build the issue graph
-	for _, issue := range childIssues {
-		node := &IssueNode{
-			ID:           issue.ID,
-			Title:        issue.Title,
-			Status:       string(issue.Status),
-			Priority:     issue.Priority,
-			DependsOn:    []string{},
-			DependedOnBy: []string{},
-			Wave:         -1, // Will be set later
-		}
-		analysis.Issues[issue.ID] = node
-
-		if issue.Status == types.StatusClosed {
-			analysis.ClosedIssues++
-		}
-	}
-
-	// Build dependency relationships (only within the epic's children)
-	childIDSet := make(map[string]bool)
-	for _, issue := range childIssues {
-		childIDSet[issue.ID] = true
-	}
-
-	for _, issue := range childIssues {
-		deps, err := s.GetDependencyRecords(ctx, issue.ID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get dependencies for %s: %w", issue.ID, err)
-		}
-
-		node := analysis.Issues[issue.ID]
-		for _, dep := range deps {
-			// Only consider dependencies within the epic (not parent-child to epic itself)
-			if dep.DependsOnID == epic.ID && dep.Type == types.DepParentChild {
-				continue // Skip the parent relationship to the epic
-			}
-			// Only track blocking dependencies
-			if !dep.Type.AffectsReadyWork() {
-				continue
-			}
-			// Only track dependencies within the epic's children
-			if childIDSet[dep.DependsOnID] {
-				node.DependsOn = append(node.DependsOn, dep.DependsOnID)
-				if targetNode, ok := analysis.Issues[dep.DependsOnID]; ok {
-					targetNode.DependedOnBy = append(targetNode.DependedOnBy, issue.ID)
-				}
-			}
-			// External dependencies to issues outside the epic
-			if !childIDSet[dep.DependsOnID] && dep.DependsOnID != epic.ID {
-				// Check if it's an external ref
-				if strings.HasPrefix(dep.DependsOnID, "external:") {
-					analysis.Warnings = append(analysis.Warnings,
-						fmt.Sprintf("%s has external dependency: %s", issue.ID, dep.DependsOnID))
-				} else {
-					analysis.Warnings = append(analysis.Warnings,
-						fmt.Sprintf("%s depends on %s (outside epic)", issue.ID, dep.DependsOnID))
-				}
-			}
-		}
-	}
-
-	// Detect structural issues
-	detectStructuralIssues(analysis, childIssues)
-
-	// Compute ready fronts (waves of parallel work)
-	computeReadyFronts(analysis)
-
-	// Set swarmable based on errors
-	analysis.Swarmable = len(analysis.Errors) == 0
-
-	return analysis, nil
-}
-
-// detectStructuralIssues looks for common problems in the dependency graph.
-//
-//nolint:unparam // issues reserved for future use
-func detectStructuralIssues(analysis *SwarmAnalysis, _ []*types.Issue) {
-	// 1. Find roots (issues with no dependencies within the epic)
-	//    These are the starting points. Having multiple roots is normal.
-	var roots []string
-	for id, node := range analysis.Issues {
-		if len(node.DependsOn) == 0 {
-			roots = append(roots, id)
-		}
-	}
-
-	// 2. Find leaves (issues that nothing depends on within the epic)
-	//    Multiple leaves might indicate missing dependencies or just multiple end points.
-	var leaves []string
-	for id, node := range analysis.Issues {
-		if len(node.DependedOnBy) == 0 {
-			leaves = append(leaves, id)
-		}
-	}
-
-	// 3. Detect potential dependency inversions
-	//    Heuristic: If a "foundation" or "setup" issue has no dependents, it might be inverted.
-	//    Heuristic: If an "integration" or "final" issue depends on nothing, it might be inverted.
-	for id, node := range analysis.Issues {
-		lowerTitle := strings.ToLower(node.Title)
-
-		// Foundation-like issues should have dependents
-		if len(node.DependedOnBy) == 0 {
-			if strings.Contains(lowerTitle, "foundation") ||
-				strings.Contains(lowerTitle, "setup") ||
-				strings.Contains(lowerTitle, "base") ||
-				strings.Contains(lowerTitle, "core") {
-				analysis.Warnings = append(analysis.Warnings,
-					fmt.Sprintf("%s (%s) has no dependents - should other issues depend on it?",
-						id, node.Title))
-			}
-		}
-
-		// Integration-like issues should have dependencies
-		if len(node.DependsOn) == 0 {
-			if strings.Contains(lowerTitle, "integration") ||
-				strings.Contains(lowerTitle, "final") ||
-				strings.Contains(lowerTitle, "test") {
-				analysis.Warnings = append(analysis.Warnings,
-					fmt.Sprintf("%s (%s) has no dependencies - should it depend on implementation?",
-						id, node.Title))
-			}
-		}
-	}
-
-	// 4. Check for disconnected subgraphs
-	// Start from roots and see if we can reach all nodes
-	visited := make(map[string]bool)
-	var dfs func(id string)
-	dfs = func(id string) {
-		if visited[id] {
-			return
-		}
-		visited[id] = true
-		if node, ok := analysis.Issues[id]; ok {
-			for _, depID := range node.DependedOnBy {
-				dfs(depID)
-			}
-		}
-	}
-
-	// Visit from all roots
-	for _, root := range roots {
-		dfs(root)
-	}
-
-	// Check for unvisited nodes (disconnected from roots)
-	var disconnected []string
-	for id := range analysis.Issues {
-		if !visited[id] {
-			disconnected = append(disconnected, id)
-		}
-	}
-
-	if len(disconnected) > 0 {
-		analysis.Warnings = append(analysis.Warnings,
-			fmt.Sprintf("Disconnected issues (not reachable from roots): %v", disconnected))
-	}
-
-	// 5. Detect cycles using simple DFS
-	// (The main DetectCycles in storage is more sophisticated, but we do a simple check here)
-	inProgress := make(map[string]bool)
-	completed := make(map[string]bool)
-	var cyclePath []string
-	hasCycle := false
-
-	var detectCycle func(id string) bool
-	detectCycle = func(id string) bool {
-		if completed[id] {
-			return false
-		}
-		if inProgress[id] {
-			hasCycle = true
-			return true
-		}
-		inProgress[id] = true
-		cyclePath = append(cyclePath, id)
-
-		if node, ok := analysis.Issues[id]; ok {
-			for _, depID := range node.DependsOn {
-				if detectCycle(depID) {
-					return true
-				}
-			}
-		}
-
-		cyclePath = cyclePath[:len(cyclePath)-1]
-		inProgress[id] = false
-		completed[id] = true
-		return false
-	}
-
-	for id := range analysis.Issues {
-		if !completed[id] {
-			if detectCycle(id) {
-				break
-			}
-		}
-	}
-
-	if hasCycle {
-		analysis.Errors = append(analysis.Errors,
-			fmt.Sprintf("Dependency cycle detected involving: %v", cyclePath))
-	}
-}
-
-// computeReadyFronts calculates the waves of parallel work.
-func computeReadyFronts(analysis *SwarmAnalysis) {
-	if len(analysis.Errors) > 0 {
-		// Can't compute ready fronts if there are cycles
-		return
-	}
-
-	// Use Kahn's algorithm for topological sort with level tracking
-	inDegree := make(map[string]int)
-	for id, node := range analysis.Issues {
-		inDegree[id] = len(node.DependsOn)
-	}
-
-	// Start with all nodes that have no dependencies (wave 0)
-	var currentWave []string
-	for id, degree := range inDegree {
-		if degree == 0 {
-			currentWave = append(currentWave, id)
-			analysis.Issues[id].Wave = 0
-		}
-	}
-
-	wave := 0
-	for len(currentWave) > 0 {
-		// Sort for deterministic output
-		sort.Strings(currentWave)
-
-		// Build titles for this wave
-		var titles []string
-		for _, id := range currentWave {
-			if node, ok := analysis.Issues[id]; ok {
-				titles = append(titles, node.Title)
-			}
-		}
-
-		front := ReadyFront{
-			Wave:   wave,
-			Issues: currentWave,
-			Titles: titles,
-		}
-		analysis.ReadyFronts = append(analysis.ReadyFronts, front)
-
-		// Track max parallelism
-		if len(currentWave) > analysis.MaxParallelism {
-			analysis.MaxParallelism = len(currentWave)
-		}
-
-		// Find next wave
-		var nextWave []string
-		for _, id := range currentWave {
-			if node, ok := analysis.Issues[id]; ok {
-				for _, dependentID := range node.DependedOnBy {
-					inDegree[dependentID]--
-					if inDegree[dependentID] == 0 {
-						nextWave = append(nextWave, dependentID)
-						analysis.Issues[dependentID].Wave = wave + 1
-					}
-				}
-			}
-		}
-
-		currentWave = nextWave
-		wave++
-	}
-
-	// Estimated sessions = total issues (each issue is roughly one session)
-	analysis.EstimatedSessions = analysis.TotalIssues
 }
 
 // renderSwarmAnalysis outputs human-readable analysis.
@@ -519,7 +138,6 @@ func renderSwarmAnalysis(analysis *SwarmAnalysis) {
 		return
 	}
 
-	// Ready fronts
 	if len(analysis.ReadyFronts) > 0 {
 		fmt.Printf("\n%s Ready Fronts (waves of parallel work):\n", ui.RenderPass("📊"))
 		for _, front := range analysis.ReadyFronts {
@@ -534,13 +152,11 @@ func renderSwarmAnalysis(analysis *SwarmAnalysis) {
 		}
 	}
 
-	// Summary stats
 	fmt.Printf("\n%s Summary:\n", ui.RenderAccent("📈"))
 	fmt.Printf("   Estimated worker-sessions: %d\n", analysis.EstimatedSessions)
 	fmt.Printf("   Max parallelism: %d\n", analysis.MaxParallelism)
 	fmt.Printf("   Total waves: %d\n", len(analysis.ReadyFronts))
 
-	// Warnings
 	if len(analysis.Warnings) > 0 {
 		fmt.Printf("\n%s Warnings:\n", ui.RenderWarn("⚠"))
 		for _, warning := range analysis.Warnings {
@@ -548,7 +164,6 @@ func renderSwarmAnalysis(analysis *SwarmAnalysis) {
 		}
 	}
 
-	// Errors
 	if len(analysis.Errors) > 0 {
 		fmt.Printf("\n%s Errors:\n", ui.RenderFail("❌"))
 		for _, err := range analysis.Errors {
@@ -556,37 +171,12 @@ func renderSwarmAnalysis(analysis *SwarmAnalysis) {
 		}
 	}
 
-	// Final verdict
 	fmt.Println()
 	if analysis.Swarmable {
 		fmt.Printf("%s Swarmable: YES\n\n", ui.RenderPass("✓"))
 	} else {
 		fmt.Printf("%s Swarmable: NO (fix errors first)\n\n", ui.RenderFail("✗"))
 	}
-}
-
-// SwarmStatus holds the current status of a swarm (computed from beads).
-type SwarmStatus struct {
-	EpicID       string        `json:"epic_id"`
-	EpicTitle    string        `json:"epic_title"`
-	TotalIssues  int           `json:"total_issues"`
-	Completed    []StatusIssue `json:"completed"`
-	Active       []StatusIssue `json:"active"`
-	Ready        []StatusIssue `json:"ready"`
-	Blocked      []StatusIssue `json:"blocked"`
-	Progress     float64       `json:"progress_percent"`
-	ActiveCount  int           `json:"active_count"`
-	ReadyCount   int           `json:"ready_count"`
-	BlockedCount int           `json:"blocked_count"`
-}
-
-// StatusIssue represents an issue in swarm status output.
-type StatusIssue struct {
-	ID        string   `json:"id"`
-	Title     string   `json:"title"`
-	Assignee  string   `json:"assignee,omitempty"`
-	BlockedBy []string `json:"blocked_by,omitempty"`
-	ClosedAt  string   `json:"closed_at,omitempty"`
 }
 
 var swarmStatusCmd = &cobra.Command{
@@ -615,18 +205,15 @@ Examples:
 	Run: func(cmd *cobra.Command, args []string) {
 		ctx := rootCtx
 
-		// Swarm commands require direct store access
 		if store == nil {
 			FatalErrorRespectJSON("no database connection")
 		}
 
-		// Resolve ID
 		issueID, err := utils.ResolvePartialID(ctx, store, args[0])
 		if err != nil {
 			FatalErrorRespectJSON("issue '%s' not found: %v", args[0], err)
 		}
 
-		// Get the issue
 		issue, err := store.GetIssue(ctx, issueID)
 		if err != nil {
 			if errors.Is(err, storage.ErrNotFound) {
@@ -637,9 +224,7 @@ Examples:
 
 		var epic *types.Issue
 
-		// Check if it's a swarm molecule - if so, follow the link to the epic
 		if issue.IssueType == "molecule" && issue.MolType == types.MolTypeSwarm {
-			// Find linked epic via relates-to dependency
 			deps, err := store.GetDependencyRecords(ctx, issue.ID)
 			if err != nil {
 				FatalErrorRespectJSON("failed to get swarm dependencies: %v", err)
@@ -662,7 +247,6 @@ Examples:
 			FatalErrorRespectJSON("'%s' is not an epic or swarm molecule (type: %s)", issueID, issue.IssueType)
 		}
 
-		// Get swarm status
 		status, err := getSwarmStatus(ctx, store, epic)
 		if err != nil {
 			FatalErrorRespectJSON("failed to get swarm status: %v", err)
@@ -673,129 +257,14 @@ Examples:
 			return
 		}
 
-		// Human-readable output
 		renderSwarmStatus(status)
 	},
-}
-
-// getSwarmStatus computes current swarm status from beads.
-func getSwarmStatus(ctx context.Context, s SwarmStorage, epic *types.Issue) (*SwarmStatus, error) {
-	status := &SwarmStatus{
-		EpicID:    epic.ID,
-		EpicTitle: epic.Title,
-		Completed: []StatusIssue{},
-		Active:    []StatusIssue{},
-		Ready:     []StatusIssue{},
-		Blocked:   []StatusIssue{},
-	}
-
-	// Get all child issues of the epic
-	childIssues, err := getEpicChildren(ctx, s, epic.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	status.TotalIssues = len(childIssues)
-	if len(childIssues) == 0 {
-		return status, nil
-	}
-
-	// Build set of child IDs for filtering
-	childIDSet := make(map[string]bool)
-	for _, issue := range childIssues {
-		childIDSet[issue.ID] = true
-	}
-
-	// Build dependency map (within epic children only)
-	dependsOn := make(map[string][]string)
-	for _, issue := range childIssues {
-		deps, err := s.GetDependencyRecords(ctx, issue.ID)
-		if err != nil {
-			continue
-		}
-		for _, dep := range deps {
-			// Skip parent-child to epic itself
-			if dep.DependsOnID == epic.ID && dep.Type == types.DepParentChild {
-				continue
-			}
-			// Only track blocking dependencies within children
-			if !dep.Type.AffectsReadyWork() {
-				continue
-			}
-			if childIDSet[dep.DependsOnID] {
-				dependsOn[issue.ID] = append(dependsOn[issue.ID], dep.DependsOnID)
-			}
-		}
-	}
-
-	// Categorize each issue
-	for _, issue := range childIssues {
-		si := StatusIssue{
-			ID:       issue.ID,
-			Title:    issue.Title,
-			Assignee: issue.Assignee,
-		}
-
-		switch issue.Status {
-		case types.StatusClosed:
-			if issue.ClosedAt != nil {
-				si.ClosedAt = issue.ClosedAt.Format("2006-01-02 15:04")
-			}
-			status.Completed = append(status.Completed, si)
-
-		case types.StatusInProgress:
-			status.Active = append(status.Active, si)
-
-		default: // open or other
-			// Check if blocked by open dependencies
-			deps := dependsOn[issue.ID]
-			var blockers []string
-			for _, depID := range deps {
-				depIssue, _ := s.GetIssue(ctx, depID)
-				if depIssue != nil && depIssue.Status != types.StatusClosed {
-					blockers = append(blockers, depID)
-				}
-			}
-
-			if len(blockers) > 0 {
-				si.BlockedBy = blockers
-				status.Blocked = append(status.Blocked, si)
-			} else {
-				status.Ready = append(status.Ready, si)
-			}
-		}
-	}
-
-	// Sort each category by ID for consistent output
-	sort.Slice(status.Completed, func(i, j int) bool {
-		return status.Completed[i].ID < status.Completed[j].ID
-	})
-	sort.Slice(status.Active, func(i, j int) bool {
-		return status.Active[i].ID < status.Active[j].ID
-	})
-	sort.Slice(status.Ready, func(i, j int) bool {
-		return status.Ready[i].ID < status.Ready[j].ID
-	})
-	sort.Slice(status.Blocked, func(i, j int) bool {
-		return status.Blocked[i].ID < status.Blocked[j].ID
-	})
-
-	// Compute counts and progress
-	status.ActiveCount = len(status.Active)
-	status.ReadyCount = len(status.Ready)
-	status.BlockedCount = len(status.Blocked)
-	if status.TotalIssues > 0 {
-		status.Progress = float64(len(status.Completed)) / float64(status.TotalIssues) * 100
-	}
-
-	return status, nil
 }
 
 // renderSwarmStatus outputs human-readable swarm status.
 func renderSwarmStatus(status *SwarmStatus) {
 	fmt.Printf("\n%s Ready Front Analysis: %s\n\n", ui.RenderAccent("🐝"), status.EpicTitle)
 
-	// Completed
 	fmt.Printf("Completed:     ")
 	if len(status.Completed) == 0 {
 		fmt.Printf("(none)\n")
@@ -808,7 +277,6 @@ func renderSwarmStatus(status *SwarmStatus) {
 		}
 	}
 
-	// Active
 	fmt.Printf("Active:        ")
 	if len(status.Active) == 0 {
 		fmt.Printf("(none)\n")
@@ -824,11 +292,9 @@ func renderSwarmStatus(status *SwarmStatus) {
 		fmt.Printf("%s\n", strings.Join(parts, ", "))
 	}
 
-	// Ready
 	fmt.Printf("Ready:         ")
 	if len(status.Ready) == 0 {
 		if len(status.Blocked) > 0 {
-			// Find what's blocking
 			needed := make(map[string]bool)
 			for _, b := range status.Blocked {
 				for _, dep := range b.BlockedBy {
@@ -852,7 +318,6 @@ func renderSwarmStatus(status *SwarmStatus) {
 		fmt.Printf("%s\n", strings.Join(parts, ", "))
 	}
 
-	// Blocked
 	fmt.Printf("Blocked:       ")
 	if len(status.Blocked) == 0 {
 		fmt.Printf("(none)\n")
@@ -866,7 +331,6 @@ func renderSwarmStatus(status *SwarmStatus) {
 		}
 	}
 
-	// Progress summary
 	fmt.Printf("\nProgress: %d/%d complete", len(status.Completed), status.TotalIssues)
 	if status.ActiveCount > 0 {
 		fmt.Printf(", %d/%d active", status.ActiveCount, status.TotalIssues)
@@ -900,18 +364,15 @@ Examples:
 		coordinator, _ := cmd.Flags().GetString("coordinator")
 		force, _ := cmd.Flags().GetBool("force")
 
-		// Swarm commands require direct store access
 		if store == nil {
 			FatalErrorRespectJSON("no database connection")
 		}
 
-		// Resolve the input ID
 		inputID, err := utils.ResolvePartialID(ctx, store, args[0])
 		if err != nil {
 			FatalErrorRespectJSON("issue '%s' not found: %v", args[0], err)
 		}
 
-		// Get the issue
 		issue, err := store.GetIssue(ctx, inputID)
 		if err != nil {
 			if errors.Is(err, storage.ErrNotFound) {
@@ -923,12 +384,10 @@ Examples:
 		var epicID string
 		var epicTitle string
 
-		// Check if it's an epic or single issue that needs wrapping
 		if issue.IssueType == types.TypeEpic || issue.IssueType == "molecule" {
 			epicID = issue.ID
 			epicTitle = issue.Title
 		} else {
-			// Auto-wrap: create an epic with this issue as child
 			if !jsonOutput {
 				fmt.Printf("Auto-wrapping single issue as epic...\n")
 			}
@@ -946,7 +405,6 @@ Examples:
 				FatalErrorRespectJSON("failed to create wrapper epic: %v", err)
 			}
 
-			// Add parent-child dependency: issue depends on epic (epic is parent)
 			dep := &types.Dependency{
 				IssueID:     issue.ID,
 				DependsOnID: wrapperEpic.ID,
@@ -965,7 +423,6 @@ Examples:
 			}
 		}
 
-		// Check for existing swarm molecule
 		existingSwarm, err := findExistingSwarm(ctx, store, epicID)
 		if err != nil {
 			FatalErrorRespectJSON("failed to check for existing swarm: %v", err)
@@ -984,7 +441,6 @@ Examples:
 			os.Exit(1)
 		}
 
-		// Validate the epic structure
 		epic, err := store.GetIssue(ctx, epicID)
 		if err != nil {
 			FatalErrorRespectJSON("failed to get epic: %v", err)
@@ -1010,7 +466,6 @@ Examples:
 			os.Exit(1)
 		}
 
-		// Create the swarm molecule
 		swarmMol := &types.Issue{
 			Title:       fmt.Sprintf("Swarm: %s", epicTitle),
 			Description: fmt.Sprintf("Swarm molecule orchestrating epic %s.\n\nEpic: %s\nCoordinator: %s", epicID, epicID, coordinator),
@@ -1026,7 +481,6 @@ Examples:
 			FatalErrorRespectJSON("failed to create swarm molecule: %v", err)
 		}
 
-		// Link swarm molecule to epic with relates-to dependency
 		dep := &types.Dependency{
 			IssueID:     swarmMol.ID,
 			DependsOnID: epicID,
@@ -1073,12 +527,10 @@ Examples:
 	Run: func(cmd *cobra.Command, args []string) {
 		ctx := rootCtx
 
-		// Swarm commands require direct store access
 		if store == nil {
 			FatalErrorRespectJSON("no database connection")
 		}
 
-		// Query for all swarm molecules
 		swarmType := types.MolTypeSwarm
 		filter := types.IssueFilter{
 			MolType: &swarmType,
@@ -1097,7 +549,6 @@ Examples:
 			return
 		}
 
-		// Build output with status for each swarm
 		type SwarmListItem struct {
 			ID          string  `json:"id"`
 			Title       string  `json:"title"`
@@ -1112,24 +563,22 @@ Examples:
 		}
 
 		var items []SwarmListItem
-		for _, swarm := range swarms {
+		for _, s := range swarms {
 			item := SwarmListItem{
-				ID:          swarm.ID,
-				Title:       swarm.Title,
-				Status:      string(swarm.Status),
-				Coordinator: swarm.Assignee,
+				ID:          s.ID,
+				Title:       s.Title,
+				Status:      string(s.Status),
+				Coordinator: s.Assignee,
 			}
 
-			// Find linked epic via relates-to dependency
-			deps, err := store.GetDependencyRecords(ctx, swarm.ID)
+			depRecords, err := store.GetDependencyRecords(ctx, s.ID)
 			if err == nil {
-				for _, dep := range deps {
+				for _, dep := range depRecords {
 					if dep.Type == types.DepRelatesTo {
 						item.EpicID = dep.DependsOnID
 						epic, err := store.GetIssue(ctx, dep.DependsOnID)
 						if err == nil && epic != nil {
 							item.EpicTitle = epic.Title
-							// Get swarm status for this epic
 							status, err := getSwarmStatus(ctx, store, epic)
 							if err == nil {
 								item.Total = status.TotalIssues
@@ -1151,10 +600,8 @@ Examples:
 			return
 		}
 
-		// Human-readable output
 		fmt.Printf("\n%s Active Swarms (%d)\n\n", ui.RenderAccent("🐝"), len(items))
 		for _, item := range items {
-			// Progress indicator
 			progressStr := fmt.Sprintf("%d/%d", item.Completed, item.Total)
 			if item.Active > 0 {
 				progressStr += fmt.Sprintf(", %d active", item.Active)
